rename log package to logger. which better reflects its purpose. Remove duplicated log.go in the project root directory, adjust references accordingly. Remove unsed randPort() method
