+++ logger/log.go
@@ -1,7 +1,7 @@
-package log
+package logger
 
 import (
-	"io/ioutil"
+	"io"
 	"os"
 	"strings"
 	"sync"
@@ -18,7 +18,7 @@ func InitializeSAM3Logger() {
 	once.Do(func() {
 		log = logrus.New()
 		// We do not want to log by default
-		log.SetOutput(ioutil.Discard)
+		log.SetOutput(io.Discard)
 		log.SetLevel(logrus.PanicLevel)
 		// Check if DEBUG_I2P is set
 		if logLevel := os.Getenv("DEBUG_I2P"); logLevel != "" {
