+++ raw/log.go
@@ -1,6 +1,6 @@
 package raw
 
-import logger "github.com/go-i2p/go-sam-go/log"
+import logger "github.com/go-i2p/go-sam-go/logger"
 
 var log = logger.GetSAM3Logger()
